### 2022-07-18 #4577 @editorialbot
[PRE REVIEW]: Choco-solver: A Java library for constraint programming
**Submitting author:** @cprudhom (<a href="http://orcid.org/0000-0002-4546-9027">Charles Prud&#39;homme</a>)
**Repository:** https://github.com/chocoteam/choco-solver
**Branch with paper.md** (empty if default branch):
**Version:** 4.10.8
