### 2022-07-29 6c832c6
implement equals and hashcode based on ID
should speed up the usage of variables within sets and maps
