### 2022-07-29 bd35dcd
hashCode returns ID directly
