### 2022-07-29 #4577 @Kevin-Mattheus-Moerman
@cprudhom thanks for this submission. Can you please check those potentially missing/invalid DOIs? :point_up: You can edit your bib file and paper where needed and then call `@editorialbot check references` to check the DOIs again, and call `@editorialbot generate pdf` to update the draft paper.
