### 2022-07-29 #4577 @jbytecode
@cprudhom - could you please correct and add the invalid and the missing DOIs as suggested?
