### 2022-07-29 #4577 @jbytecode
Dear @cprudhom,

It seems

- the paper has two authors. it is clear that the most of the contributions were handled by the authors.
- there are more commits from other contributors including @dimitri-justeau (161 commits), @ArthurGodet (39 commits), @JLiangWaterloo (36 commits), @xlorca, @xela85, etc.

could you please clarify why those people are omitted from the manuscript as authors.

Thank you in advance.
