### 2022-08-16 #4577 @cprudhom
Dear @jbytecode,

Thank you for reviewing the article. I was out of the office but this week I should be able to look at the comments and get back asap.

Best regards
