### 2022-08-16 #4577 @cprudhom
> It seems
>
> * the paper has two authors. it is clear that the most of the contributions were handled by the authors.
> * there are more commits from other contributors including @dimitri-justeau (161 commits), @ArthurGodet (39 commits), @JLiangWaterloo (36 commits), @xlorca, @xela85, etc.
>
> could you please clarify why those people are omitted from the manuscript as authors.

@jbytecode I made the choice the define as authors the main contributors of the library, from design to implementation.
The other contributors you mentioned (and I think there are some missing) have have participated in the improvement of the library, either by correcting bugs or by enriching it, to a lesser extent. I have chosen to mention their names in the Acknowledgements section, but I suppose that can be modified.
