### 2022-08-22 70764c3
remove dead scripts + update headers
