+++ solver/bin/package.sh
@@ -1,118 +1,0 @@
-#!/bin/bash
-
-source ./commons.sh
-set -ex
-#
-# Make archive for the release
-function makeArchive(){
-    VERSION=$1
-    if [ -d choco-${VERSION} ]; then
-      rm -Rf choco-${VERSION};
-    fi
-    mkdir choco-${VERSION} || quit "unable to make choco-${VERSION}"
-    git checkout ${VERSION} || quit "unable to checkout ${VERSION}"
-    mvn -q clean install -DskipTests || quit "unable to install"
-
-    mv ./target/choco-solver-${VERSION}.jar ./choco-${VERSION} || quit "unable to mv jar"
-    mv ./target/choco-solver-${VERSION}-no-dep.jar ./choco-${VERSION}|| quit "unable to mv no-dep"
-    mv ./target/choco-solver-${VERSION}-sources.jar ./choco-${VERSION}|| quit "unable to mv src"
-
-    cp ./user_guide.pdf ./choco-${VERSION}/user_guide-${VERSION}.pdf|| quit "unable to cp user guide"
-    cp ./README.md ./choco-${VERSION}|| quit "unable to cp README"
-    cp ./CHANGES.md ./choco-${VERSION}|| quit "unable to cp CHANGES"
-    cp ./LICENSE ./choco-${VERSION}|| quit "unable to cp LICENSE"
-
-    mvn -q javadoc:aggregate  || quit "unable to javadoc"
-    cd target/site/|| quit "unable to cd target/site"
-    zip -r apidocs-${VERSION}.zip ./apidocs/*|| quit "unable to zip apidocs"
-    mv apidocs-${VERSION}.zip ../../choco-${VERSION}|| quit "unable to mv apidocs"
-    cd ../../|| quit "unable to cd ../../"
-
-    zip choco-${VERSION}.zip ./choco-${VERSION}/*|| quit "unable to zip dir"
-}
-#
-# push javadoc onto website
-function pushJavadoc(){
-    VERSION=$1
-    LOCAL=`mktemp -d -t choco.XXX`|| quit "unable to make temp dir"
-    echo ${LOCAL}
-    git -C ${LOCAL} init
-    git -C ${LOCAL} remote add origin "https://github.com/chocoteam/chocoteam.github.io.git"||exit 1
-    git -C ${LOCAL} fetch origin||exit 1
-    git -C ${LOCAL} checkout master||git -C ${LOCAL} checkout -b master
-    unzip -o ./choco-${VERSION}/apidocs-${VERSION}.zip -d ${LOCAL}|| quit "unable to unzip apidocs"
-    cd ${LOCAL}|| quit "unable to go into website"
-    git add ./apidocs/ || quit "unable to add apidocs"
-    git commit -a -m "push javadoc of ${VERSION}"|| quit "unable to commit"
-    git push origin master || quit "unable to push website"
-    cd -
-    rm -rf ${LOCAL} || quit "unable to remove ${LOCAL}"
-}
-#
-# extract comment from CHANGES.md
-function extractReleaseComment(){
-    VERSION=$1
-    LINES=$(grep -ne "-------------------" CHANGES.md |  cut -d : -f1 | tr "\n" ",")
-    IFS=',' read -r -a ARRAY <<< "$LINES"
-    FROM=$(expr ${ARRAY[0]} + 1)
-    TO=$(expr ${ARRAY[1]} - 2)
-    CHANGES=$(cat ./CHANGES.md | sed -n "${FROM},${TO}p" | perl -p -e 's/  /\\t/' | perl -p -e 's/(\#)([0-9])/\\#\2/' | perl -p -e 's/\n/\\n/g')
[…]
