### 2022-08-22 9adb7b8
Update of the preferred contact medium
