+++ scripts/release.sh
@@ -29,6 +29,10 @@ git ls-remote --exit-code --tags origin ${VERSION} && quit "tag ${VERSION} alrea
 git tag -a ${VERSION} -m "create tag ${VERSION}" || quit "Unable to tag with ${VERSION}"
 git push --tags || quit "Unable to push the tag ${VERSION}"
 
+
+# Proceed to the deployment
+# mvn -P ossrhDeploy  javadoc:jar source:jar deploy -DskipTests -B -U  ||quit "Unable to deploy to master"
+
 #Set the next development version
 echo "** Prepare master for the next version **"
 ./scripts/set_version.sh --next ${VERSION}
@@ -37,45 +41,45 @@ git commit -m "Prepare the code for the next version" -a ||quit "Unable to commi
 #Push changes on develop, with the tag
 git push origin master ||quit "Unable to push to master"
 
-# Package the current version
-GH_API="https://api.github.com/repos/chocoteam/choco-solver/"
-GH_UPL="https://uploads.github.com/repos/chocoteam/choco-solver/"
-
-AUTH="Authorization: token ${GH_TOKEN}"
-
-# Validate token.
-curl -o /dev/null -i -sH "${AUTH}" "${GH_API}releases" || quit "Error: Invalid repo, token or network issue!";
-
-# prepare release comment
-
-#find position of release separator in CHANGES.md, only keep the 2nd and 3rd
-temp_file="tmpreadme.json"
-$(touch ${temp_file}) || quit "Unable to create tmp file"
-
-# extract release comment
-extractReleaseComment ${VERSION} ${temp_file} || quit "Unable to extract release comment"
-
-# create release
-response=$(curl -i -sH "$AUTH" --data @${temp_file} "${GH_API}releases") || quit "Unable to create the release"
-
-# get the asset id
-ID=$(echo "$response" | grep -m 1 "id.:"| tr : = | tr -cd '[[:alnum:]]=') || quit "Error: Failed to get release id for tag: ${VERSION}"; echo "$response" | awk 'length($0)<100' >&2
-ID=(${ID//=/ }) || quit "Error: Unable to split id: ${ID}"
-id=${ID[1]} || quit "Error: Unable to get id: ${ID}"
-
-# add asset
-curl -i -sH "$AUTH" -H "Content-Type: application/zip" \
-         -data-binary @choco-${VERSION}.zip \
-         "${GH_UPL}/releases/${id}/assets?name=choco-${VERSION}.zip" \
-         || quit "Unable to add asset"
-
-
-# create the next milestone
-NEXT=$(echo "${VERSION%.*}.$((${VERSION##*.}+1))") || quit "Unable to get next release number"
-curl -i -sH "$AUTH" --data '{ "title": '\""${NEXT}"\"'}' "${GH_API}milestones"
-
-if [ -d choco-${VERSION} ]; then
-  rm -Rf choco-${VERSION} || quit "Unable to remove choco-${VERSION} dir"
-fi
-#rm choco-${VERSION}.zip
-rm ${temp_file} || quit "Unable to remove tmp file"
+## Package the current version
+#GH_API="https://api.github.com/repos/chocoteam/choco-solver/"
+#GH_UPL="https://uploads.github.com/repos/chocoteam/choco-solver/"
+#
+#AUTH="Authorization: token ${GH_TOKEN}"
+#
+## Validate token.
+#curl -o /dev/null -i -sH "${AUTH}" "${GH_API}releases" || quit "Error: Invalid repo, token or network issue!";
+#
+## prepare release comment
+#
+##find position of release separator in CHANGES.md, only keep the 2nd and 3rd
[…]
