### 2022-08-22 #4577 @jbytecode
@cprudhom - do you have more suggestions? I think it is hard to find reviewers that have both experiences on Java and constraint programming. The search space can be broadened using different keywords.

If you have no more suggestions, I will try to find reviewers, of course.

Thank you in advance.
