### 2022-08-23 #4577 @skadio
Team, I can review this submission.

In fact, Choco has been a standard tool of choice in my academic teaching and research when it comes to JAVA. It is a well-designed and performant software and its contributions to the community is well-known. Kudos to all contributors for their work!

This manuscript would make a contribution to JOSS community. More broadly, it would also help bridge communities: I doubt if many CP (and by extension SAT, ILP etc.) are aware of the JOSS community.

That said, it is a busy time of the year (with moving into the last quarter after summer) but this is right at the intersection of CP and open-source so if you allow me a couple weeks, I can return a full review.

I read the manuscript and here are some immediate observations on the write-up to parallelize the work:

* The paper reads well for a CP researcher, but for JOSS community, I have a few suggestions that should be easy to make:

* The Abstract goes into a CP Overview and the link between Choco and CP is kinda gets lost and/or not clear. This might be obvious for a CP person but not necessarily to a general reader. Some of the technical details here in abstract are also "too technical", in my opinion. For example, it mixes constraints and predicates (are they the same?).

* Re: Abstract. If I wear a general user hat, it is not clear that's the value proposition. We need to answer why should the user/reader be interested using it. Choco has many great applications, why not mention a few of these industrial domains as success stories for CP/Choco? Then, the user can achieve in their domains using Choco. That "promise/proposition" shall be added to the abstract.

* We need an example CP problem to make the scenario concrete for the user. Why not consider N-Queens or sth similar (and then we can build on the CP Model). You might even make 2 models; one abstract model and then one Java Choco implementation. What we would like to highlight here is that there is no gap between problem description and the implementation. That "declarative" nature is the strength of CP/Choco.

* The example can help bridge the definitions (constraints, domains, variables) and the Features/Functionality.
*
