+++ solver/src/test/java/org/chocosolver/solver/constraints/nary/min_max/ArgmaxminTest.java
@@ -33,7 +33,17 @@ import java.util.Random;
 public class ArgmaxminTest {
 
     @DataProvider
-    private Object[][] decOrNot() {
+    private Object[][] dec() {
+        List<Object[]> args = new ArrayList<>();
+        for (int offset = -3; offset < 4; offset++) {
+            args.add(new Object[]{true, offset});
+            args.add(new Object[]{false, offset});
+        }
+        return args.toArray(new Object[0][0]);
+    }
+
+    @DataProvider
+    private Object[][] decAndSeed() {
         List<Object[]> args = new ArrayList<>();
         for (int seed = 0; seed < 200; seed++) {
             args.add(new Object[]{true, seed});
@@ -64,9 +74,9 @@ public class ArgmaxminTest {
         model.member(z, new int[]{0, 2, 3}).post();
         Solver solver = model.getSolver();
         solver.setSearch(Search.randomSearch(ArrayUtils.append(x, x, new IntVar[]{z}), seed));
-        while (solver.solve()){
-            for(int i=0;i<x.length;i++){
-                Assert.assertFalse(i!=z.getValue() && x[i].getValue()<min.getValue());
+        while (solver.solve()) {
+            for (int i = 0; i < x.length; i++) {
+                Assert.assertFalse(i != z.getValue() && x[i].getValue() < min.getValue());
             }
         }
         Assert.assertEquals(solver.getSolutionCount(), 234);
@@ -86,16 +96,16 @@ public class ArgmaxminTest {
         model.member(z, new int[]{0, 2, 3}).post();
         Solver solver = model.getSolver();
         solver.setSearch(Search.randomSearch(ArrayUtils.append(x, x, new IntVar[]{z}), seed));
-        while (solver.solve()){
-            for(int i=0;i<x.length;i++){
-                Assert.assertFalse(i!=z.getValue() && x[i].getValue()>max.getValue());
+        while (solver.solve()) {
+            for (int i = 0; i < x.length; i++) {
+                Assert.assertFalse(i != z.getValue() && x[i].getValue() > max.getValue());
             }
         }
         Assert.assertEquals(solver.getSolutionCount(), 84);
         solver.printShortStatistics();
     }
 
-    @Test(groups = "1s", dataProvider = "decOrNot")
+    @Test(groups = "1s", dataProvider = "decAndSeed")
     public void test1(boolean dec, long seed) {
         Model model = new Model(Settings.init());
         IntVar[] x = new IntVar[4];
@@ -116,7 +126,7 @@ public class ArgmaxminTest {
         solver.printShortStatistics();
     }
 
-    @Test(groups = "1s", dataProvider = "decOrNot")
+    @Test(groups = "1s", dataProvider = "decAndSeed")
     public void test1o(boolean dec, long seed) {
         Model model = new Model(Settings.init());
         IntVar[] x = new IntVar[4];
@@ -133,7 +143,29 @@ public class ArgmaxminTest {
         Solver solver = model.getSolver();
         solver.setSearch(Search.randomSearch(ArrayUtils.append(x, x, new IntVar[]{z}), seed));
         solver.findAllSolutions();
-        Assert.assertEquals(solver.getSolutionCount(), 24);
+        Assert.assertEquals(solver.getSolutionCount(), 84);
+        solver.printShortStatistics();
+    }
+
+    @Test(groups = "1s", dataProvider = "decAndSeed")
+    public void test1o2(boolean dec, long seed) {
+        Model model = new Model(Settings.init());
[…]
