### 2022-08-24 c0c3605
Add Ir to argmax (#929)
