+++ solver/src/main/java/org/chocosolver/solver/constraints/nary/min_max/PropArgmax.java
@@ -9,6 +9,7 @@
  */
 package org.chocosolver.solver.constraints.nary.min_max;
 
+import org.chocosolver.memory.IStateBitSet;
 import org.chocosolver.memory.IStateInt;
 import org.chocosolver.solver.constraints.Propagator;
 import org.chocosolver.solver.constraints.PropagatorPriority;
@@ -41,6 +42,7 @@ public class PropArgmax extends Propagator<IntVar> {
     private final int o;
     private final IStateInt ubi;
     private final IStateInt lbi;
+    private final IStateBitSet Ir;
     private final IIntDeltaMonitor delta;
     private final IntProcedure proc;
 
@@ -51,11 +53,16 @@ public class PropArgmax extends Propagator<IntVar> {
         this.delta = z.monitorDelta(this);
         this.ubi = z.getModel().getEnvironment().makeInt(-1);
         this.lbi = z.getModel().getEnvironment().makeInt(-1);
+        this.Ir = z.getModel().getEnvironment().makeBitSet(n);
         this.proc = j -> {
-            if (j - o == ubi.get()) {
+            int jj = j - o;
+            int lbi_ = lbi.get();
+            int ubi_  = ubi.get();
+            Ir.set(jj, vars[jj].getUB() > vars[lbi_].getLB() - (jj <= lbi_ ? 1 : 0));
+            if (jj == ubi_) {
                 filterUb();
             } else {
-                vars[j - o].updateUpperBound(vars[ubi.get()].getUB() - (j < lbi.get() ? 1 : 0), PropArgmax.this);
+                vars[jj].updateUpperBound(vars[ubi_].getUB() - (j < lbi_ ? 1 : 0), PropArgmax.this);
             }
         };
     }
@@ -73,7 +80,12 @@ public class PropArgmax extends Propagator<IntVar> {
     public void propagate(int evtmask) throws ContradictionException {
         vars[n].updateBounds(o, n + o - 1, this);
         filterUb();
-        lbi.set(argmaxlb(IntVar::getLB));
+        int lbi_ = argmaxlb(IntVar::getLB);
+        lbi.set(lbi_);
+        for (int j = vars[n].nextValueOut(-1 + o); j < n + o; j = vars[n].nextValueOut(j)) {
+            int jj = j - o;
+            Ir.set(jj, vars[jj].getUB() > vars[lbi_].getLB() - (jj <= lbi_ ? 1 : 0));
+        }
         filterZ();
         if (vars[n].isInstantiated()) {
             filterLb(vars[n].getValue() - o);
@@ -137,26 +149,31 @@ public class PropArgmax extends Propagator<IntVar> {
     }
 
     private void filterUb() throws ContradictionException {
-        ubi.set(argmaxub(IntVar::getUB));
-        int ub = vars[ubi.get()].getUB();
-        for (int j = vars[n].nextValueOut(-1 + o); j < n + o; j = vars[n].nextValueOut(j)) {
-            int jj = j - o;
-            vars[jj].updateUpperBound(ub - (jj < ubi.get() ? 1 : 0), this);
+        int ubi_ = argmaxub(IntVar::getUB);
+        ubi.set(ubi_);
+        int ub = vars[ubi_].getUB();
+        for (int jj = Ir.nextSetBit(0); jj > -1; jj = Ir.nextSetBit(jj + 1)) {
+            vars[jj].updateUpperBound(ub - (jj < ubi_ ? 1 : 0), this);
         }
     }
 
     private void filterZ() throws ContradictionException {
-        int lb = vars[lbi.get()].getLB();
+        int lbi_ = lbi.get();
+        int lb = vars[lbi_].getLB();
[…]
