### 2022-08-24 #4577 @jbytecode
👋👋👋 Dear @diegomcarvalho    👋👋👋

We need a second reviewer for this submission.

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**Choco-solver: A Java library for constraint programming**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4577).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a seperate thread. In that thread, there will be 20 check items for each single reviewer.

Thank you in advance!
