### 2022-08-25 #4577 @cprudhom
> I think the author @cprudhom will consider your early suggestions before we set the second reviewer.
Sure, two questions though:
- should I answer in this thread or rather in #4708 ?
- Can I modify the article to take into account the remarks (and inform the reviewers) or do I have to wait until the end of the review?
Thank you in advance
