### 2022-08-25 #4577 @jbytecode
@cprudhom

- please use the review thread, the review has been started. This thread will no longer be used.
- yes, please modify your article, generate pdf in the review thread. The review process in JOSS is interactive. Authors and reviewers always interact each other.
