### 2022-08-25 #4708 @cprudhom
@skabio

> The paper reads well for a CP researcher, but for JOSS community, I have a few suggestions that should be easy to make:
>
> The Abstract goes into a CP Overview and the link between Choco and CP is kinda gets lost and/or not clear. This might be obvious for a CP person but not necessarily to a general reader. Some of the technical details here in abstract are also "too technical", in my opinion. For example, it mixes constraints and predicates (are they the same?).

I will try to make the CP Overview easier to understand for non CP person.
I made the choice to introduce Choco in the 2nd section, but I understand that it is a bit late, right?

> Re: Abstract. If I wear a general user hat, it is not clear that's the value proposition. We need to answer why should the user/reader be interested using it. Choco has many great applications, why not mention a few of these industrial domains as success stories for CP/Choco? Then, the user can achieve in their domains using Choco. That "promise/proposition" shall be added to the abstract.

I can try emphasize the versatility of the library (well, the CP in general) through success stories and highlight the salient features of Choco.

> We need an example CP problem to make the scenario concrete for the user. Why not consider N-Queens or sth similar (and then we can build on the CP Model). You might even make 2 models; one abstract model and then one Java Choco implementation. What we would like to highlight here is that there is no gap between problem description and the implementation. That "declarative" nature is the strength of CP/Choco.

I didn't think I could go into that kind of detail, but it's an option.

> The example can help bridge the definitions (constraints, domains, variables) and the Features/Functionality.

@ozgurakgun Would you prefer to comment on the current version or wait for a new version that takes these remarks into account?
