### 2022-08-25 #4708 @skadio
@ozgurakgun to save time/effort for our authors here, would you mind a quick look at the write-up (it is 2 pages) and comment on whether you would agree/disagree with my points on the manuscript.

My main commentary is from the perspective of a general JOSS reader (think non-CP researcher) for 1) a quick rephrase of the abstract to make the value proposition of the tool (and hence CP) clearer and 2) an illustrative example to make the CP concepts and Choco functionality more concrete to raise appetite to adopt/try the tool (and CP technology)

These two comments are motivated by the great opportunity here to showcase Choco and CP to the general open-source community via JOSS.

We would not want @cprudhom to change the paper back and forth if you interpret it differently, so a quick directional comment from you can help save us time/energy.
