### 2022-08-25 #4708 @ozgurakgun
Hi @skadio and all

I have now read the text and I agree it would benefit from an example. I am afraid n-queens may be too small/trivial?

Also I am not sure how much detail you think the authors should include. Constructing a full example using Choco would take too much space and wouldn't be very interesting. Instead the authors may give some example use cases without going into too much detail? Some of these examples can be chosen from those that will be interesting to the software development community: things like dependency resolution[1], type inference[2], test case generation[3], etc.

[1] https://www.well-typed.com/blog/2008/08/solving-the-diamond-dependency-problem/
[2] https://www.microsoft.com/en-us/research/publication/type-inference-as-constraint-solving-how-ghcs-type-inference-engine-actually-works/
[3] https://link.springer.com/chapter/10.1007/978-3-642-12592-8_10

(I am sorry the first 2 links are to blog articles instead of peer-reviewed papers, but several papers in these areas exist)

In addition to this point about examples I think it would make sense to mention/cite CP modelling languages that can use Choco as a backend solver. This shows how versatile Choco can be: it can be used directly to build a model or it can be used as a black-box solver with models generated from higher-level tools.
