### 2022-08-28 #4708 @jbytecode
@cprudhom , @ozgurakgun - of course, the listed subjects are appropriate for an example, but whatever the example is, it is better to get it as small as possible in size. The main objective of providing an example in JOSS papers is to present the basic usage of the software.

And of course, the decision will be made by our reviewers.

Thank you in advance.
