### 2022-09-12 849037d
SparseBitSet for a backtrackable bitset with mostly 0s.
This implementation for IStateBitSet targets backtrackable bitsets that mostly contains 0.
For non-dense bitsets, this implementation is more efficient in terms of memory usage as bitsets are allocated per region instead of relying
on a single, possibly bit array. In this context, the iteration is also faster.
