+++ solver/src/test/java/org/chocosolver/memory/structure/SparseBitSetTest.java
@@ -1,0 +1,401 @@
+package org.chocosolver.memory.structure;
+
+import org.chocosolver.memory.IEnvironment;
+import org.chocosolver.memory.IStateBitSet;
+import org.chocosolver.memory.trailing.EnvironmentTrailing;
+import org.chocosolver.solver.Model;
+import org.testng.Assert;
+import org.testng.annotations.DataProvider;
+import org.testng.annotations.Test;
+
+import java.util.Random;
+
+/**
+ * Unit tests for {@link org.chocosolver.memory.structure.SparseBitSet}.
+ */
+public class SparseBitSetTest {
+
+  @Test
+  public void testNew() {
+    IEnvironment env = new EnvironmentTrailing();
+    IStateBitSet bs = new SparseBitSet(env, 64);
+    Assert.assertTrue(bs.isEmpty());
+    Assert.assertEquals(0, bs.cardinality());
+    Assert.assertEquals(-1, bs.nextSetBit(0));
+  }
+
+  @Test
+  public void testSetGet() {
+    IEnvironment env = new EnvironmentTrailing();
+    IStateBitSet bs = new SparseBitSet(env, 64);
+    Assert.assertFalse(bs.get(72904));
+    bs.set(65);
+    Assert.assertTrue(bs.get(65));
+    Assert.assertFalse(bs.get(64));
+    Assert.assertEquals(1, bs.cardinality());
+    bs.set(0);
+    Assert.assertEquals(2, bs.cardinality());
+    Assert.assertTrue(bs.get(0));
+
+    // Idempotence.
+    bs.set(0, true);
+    Assert.assertEquals(2, bs.cardinality());
+
+    bs.set(0, false);
+    Assert.assertEquals(1, bs.cardinality());
+    Assert.assertFalse(bs.get(0));
+  }
+
+  @Test
+  public void testNextSetBit() {
+    IEnvironment env = new EnvironmentTrailing();
+    IStateBitSet bs = new SparseBitSet(env, 64);
+    // Set inside the first block.
+    bs.set(1);
+    // Other consecutive bit sets on different blocks.
+    bs.set(255);
+    bs.set(256);
+    bs.set(257);
+
+    // Right on the spot
+    Assert.assertEquals(bs.nextSetBit(1), 1);
+    Assert.assertEquals(bs.nextSetBit(255), 255);
+    Assert.assertEquals(bs.nextSetBit(256), 256);
+    Assert.assertEquals(bs.nextSetBit(257), 257);
+
+    // Before, same block.
+    Assert.assertEquals(bs.nextSetBit(0), 1);
+    // Before previous blocks.
+    Assert.assertEquals(bs.nextSetBit(128), 255);
+    // After, used block.
+    Assert.assertEquals(bs.nextSetBit(258), -1);
+    // After, unused block.
+    Assert.assertEquals(bs.nextSetBit(65445), -1);
+  }
+
+  @Test
+  public void testNextClearBit() {
+    final IEnvironment env = new EnvironmentTrailing();
+    IStateBitSet bs = new SparseBitSet(env, 64);
+    Assert.assertEquals(0, bs.nextClearBit(0));
+    bs.set(1);
+    // Before the set bit, same block.
+    Assert.assertEquals(0, bs.nextClearBit(0));
+    // Right on the set bit. Expect the next one on this block.
+    Assert.assertEquals(2, bs.nextClearBit(1));
+
+    // Last clear bit on the same block.
+    Assert.assertEquals(63, bs.nextClearBit(63));
+
+    // On a non-existing block.
+    Assert.assertEquals(135, bs.nextClearBit(135));
+
[…]
