### 2022-09-16 38b0254
Fix lack of filtering in STR2+ when domain is bounded
