### 2022-09-16 8ab9b62
Integration comments from JGF
