### 2022-09-16 #4708 @skadio
Thank you for the revised version that improves over the initial submission. I especially like the new Summary and showcasing practical industry applications.

While I still believe that a formal Definition of a CSP/COP and then illustrative an example to make that Definition concrete would be good for the general reader, I won't hold that against the paper.

I doubt any CP-er would have difficulty reading the paper (in fact, they should know/heard of Choco anyway) so the question comes down to: "would the paper score the opportunity for a non-CP/MP/SAT person be excited about the paradigm to follow the references/links to find Modeling Tutorial. I will follow your judgement on that, and happy with the current version.

Minor comments on the write-up:
- Statement of Need: is Choco the only Java solver, if so, stating that would make the statement stronger. If not, pointing out the differences with existing ones would be scholarly.
- L29 typo the CP --> CP
- L37, _filters_ ,removes, is this a typo? why is removes in comma and/or filter in italics
- Nitpicking comment on L27-28: "Like ILP/SAT, CP is a field of MP". I kindly disagree with that statement, and many would do too, I think. Viewing SAT as MP could be problematic since its purely logical, and most CP constraints are not even numerical expressions but semantic constraints. Why not repharse to sth like: CP is a declarative paradigm where the user states the problem and the computer solves it. It is hence related to other approaches such as SAT and MP but is distinguished from them with its high-level modeling language and expressiveness. Sth like that.. Hope that makes sense.
