### 2022-09-16 #4708 @skadio
Related to my Statemenf of Need comment: the review form has a question on "Do the authors describe how this software compares to other commonly-used packages?" Something to consider.

With these minor comments, the paper is good to go from my end. Thank you @cprudhom
