### 2022-09-27 #4708 @cprudhom
I'm waiting for a paragraph from my co-author about industrial applications.
If that does not seem relevant to you, I will post-pone it and do my best to consider reviews by the end of the week
