### 2022-09-27 #4708 @ozgurakgun
Just to confirm, I have gone through the checklist and am generally happy with the text. Adding that paragraph about industrial applications will only help the paper, so assuming we are not in a rush I would vote for waiting a little bit. Thanks for doing the work for this @cprudhom - increasing Choco's (and hence CP's) visibility in the open-source software development community is definitely a good thing.
