### 2022-09-28 1ee1a22
Fix performance issue of ViewDeltaMonitor
