### 2022-09-28 46b678d
Add missing checks in OneWordS32BitSet and OneWordS64BitSet + update constructors
