### 2022-09-28 bf5a982
Remove toString from class that extends AbstractSet to avoid bug in debug mode due to nested calls to iterator
