+++ examples/src/main/java/org/chocosolver/examples/integer/Semafor.java
@@ -1,0 +1,162 @@
+/*
+ * This file is part of examples, http://choco-solver.org/
+ *
+ * Copyright (c) 2022, IMT Atlantique. All rights reserved.
+ *
+ * Licensed under the BSD 4-clause license.
+ *
+ * See LICENSE file in the project root for full license information.
+ */
+package org.chocosolver.examples.integer;
+
+import org.chocosolver.solver.Model;
+import org.chocosolver.solver.Solver;
+import org.chocosolver.solver.variables.BoolVar;
+import org.chocosolver.solver.variables.IntVar;
+import org.chocosolver.solver.variables.SetVar;
+
+import java.util.Arrays;
+import java.util.stream.IntStream;
+
+/**
+ * <br/>
+ *
+ * @author Charles Prud'homme
+ * @since 15/09/2022
+ */
+public class Semafor {
+
+    public static void main(String[] args) {
+        new Semafor();
+    }
+
+    public static final int NB_MAX_SRVS = 3;
+    public static final int NB_MAX_VMS = 3;
+    public static final int NB_MAX_CPUCORE = 24;
+    public static final int NB_MAX_RAM = 64; // in Go
+
+    private static class Server {
+        protected static int ID = 0;
+        int uuid = ID++;
+        String name;
+        IntVar cpuCore;
+        IntVar ram;
+
+        SetVar hostingVMS;
+
+        public Server(String name, Model model) {
+            this.name = name;
+            cpuCore = model.intVar(name + "_cpuCore", 0, NB_MAX_CPUCORE);
+            ram = model.intVar(name + "_ram", 0, NB_MAX_RAM);
+            // could be refined if potential VM were known at creation
+            hostingVMS = model.setVar(name + "_hosts", new int[]{}, IntStream.range(0, NB_MAX_VMS).toArray()); // 4 VMs max
+        }
+
+        public void implicitConstraints(Model model, VM... vms) {
+            IntVar[] usedCPU = model.intVarArray(NB_MAX_VMS, 0, 0);
+            IntVar[] usedRAM = model.intVarArray(NB_MAX_VMS, 0, 0);
+            BoolVar[] hostedBy = IntStream.range(0, NB_MAX_VMS).mapToObj(i -> model.boolVar(false)).toArray(BoolVar[]::new);
+            for (VM vm : vms) {
+                hostedBy[vm.uuid] = vm.hostedBy.eq(uuid).boolVar();
+                usedCPU[vm.uuid] = hostedBy[vm.uuid].mul(vm.cpuCore).intVar();
+                usedRAM[vm.uuid] = hostedBy[vm.uuid].mul(vm.ram).intVar();
+            }
+            model.sum(usedCPU, "<=", cpuCore).post();
+            model.sum(usedRAM, "<=", ram).post();
+            model.setBoolsChanneling(hostedBy, hostingVMS).post();
+        }
+
+    }
+
+    private static class VM {
+        protected static int ID = 0;
+        int uuid = ID++;
+        String name;
+        IntVar cpuCore;
+        IntVar ram;
+        IntVar hostedBy;
+
+        public VM(String name, Model model) {
+            this.name = name;
+            cpuCore = model.intVar(name + "_cpuCore", 0, NB_MAX_CPUCORE);
+            ram = model.intVar(name + "_ram", 0, NB_MAX_RAM);
+            // could be refined if potential SRV were known at creation
+            hostedBy = model.intVar(name + "_hostedBy", 0, NB_MAX_SRVS);
+        }
+
+        public void mustBeHosted(Model model, Server... srvs) {
+            model.member(hostedBy, Arrays.stream(srvs).mapToInt(v -> v.uuid).toArray()).post();
+        }
+
[…]
