### 2022-09-28 #4708 @cprudhom
@skadio

> Statement of Need: is Choco the only Java solver, if so, stating that would make the statement stronger. If not, pointing out the differences with existing ones would be scholarly.

Actually, there are (at least) two other mature constraint solvers in Java : [JaCoP](https://github.com/radsz/jacop) and [ACE (ex-Abscon)](https://github.com/xcsp3team/ace).
Both of them are very good constraint solver and their performances are quite similar to the ones of Choco but I think that Choco is more user-oriented (documentation, tutorials, frequent releases, regression and performance tests, ...).

> L29 typo the CP --> CP

👍
> L37, filters ,removes, is this a typo? why is removes in comma and/or filter in italics

I will keep filters and remove italics.

> Nitpicking comment on L27-28: "Like ILP/SAT, CP is a field of MP". I kindly disagree with that statement, and many would do too, I think. Viewing SAT as MP could be problematic since its purely logical, and most CP constraints are not even numerical expressions but semantic constraints. Why not repharse to sth like: CP is a declarative paradigm where the user states the problem and the computer solves it. It is hence related to other approaches such as SAT and MP but is distinguished from them with its high-level modeling language and expressiveness. Sth like that.. Hope that makes sense.

You are right, I will rephrase the two first sentences.
