### 2022-09-28 #4708 @skadio
Thank you @cprudhom. Appreciate the input and revisions.

It might be good for the general user to be aware of other Java solvers then, with commentary on how Choco adds value along with some constructive verbiage around on those differences. This would help people position this work with respect to the field, and more such options also prove the maturity/applicability/relevance of the field as opposed to just one single option.

Thanks!
