### 2022-10-04 023096c
first version of Model.analyse method
