+++ solver/src/main/java/org/chocosolver/solver/Model.java
@@ -27,10 +27,14 @@ import org.chocosolver.solver.objective.IObjectiveManager;
 import org.chocosolver.solver.objective.ObjectiveFactory;
 import org.chocosolver.solver.propagation.PropagationEngine;
 import org.chocosolver.solver.variables.*;
+import org.chocosolver.solver.variables.view.IView;
+import org.chocosolver.util.ESat;
 import org.chocosolver.util.tools.VariableUtils;
 
+import java.io.PrintStream;
 import java.util.*;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
@@ -1054,5 +1058,139 @@ public class Model implements IModel {
     @Override
     public Model ref() {
         return this;
+    }
+
+    private static String getClassName(Class c) {
+        String[] sp = c.toString().split("\\.");
+        return sp[sp.length - 1];
+    }
+
+    private static <V> String prettyIntListSizeMap(Map<Integer, List<V>> map) {
+        StringBuilder sb = new StringBuilder("{");
+        List<Integer> sortedArities = map.keySet().stream().sorted().collect(Collectors.toList());
+        for (int i = 0; i < sortedArities.size(); i++) {
+            sb.append(sortedArities.get(i)).append(": ").append(map.get(sortedArities.get(i)).size());
+            if (i + 1 < sortedArities.size()) {
+                sb.append(", ");
+            }
+        }
+        sb.append("}");
+        return sb.toString();
+    }
+
+    public void analyse() {
+        analyse(System.out);
+    }
+
+    public void analyse(PrintStream ps) {
+        ps.println();
+        ps.println("################################################################################################");
+        ps.println("#################################### BEGIN OF MODEL ANALYSIS ###################################");
+        ps.println("################################################################################################");
+        ps.println();
+        ps.println("################################################");
+        ps.println("############## VARIABLES ANALYSIS ##############");
+        ps.println("################################################");
+        ps.println();
+        Class[] varsTypes = new Class[]{BoolVar.class, GraphVar.class, IntVar.class, RealVar.class, SetVar.class};
+        Map<Class, Map<String, List<Variable>>> mapTypeClassNbVars = new HashMap<>();
+        for (Class c : varsTypes) {
+            mapTypeClassNbVars.put(c, Arrays.stream(vars)
+                    .filter(Objects::nonNull)
+                    .filter(c::isInstance)
+                    .filter(var -> !c.equals(IntVar.class) || !(var instanceof BoolVar))
+                    .collect(Collectors.groupingBy(var -> getClassName(var.getClass()))));
+            if (!mapTypeClassNbVars.get(c).isEmpty()) {
+                ps.println(getClassName(c));
+                List<String> list = mapTypeClassNbVars.get(c).keySet().stream().sorted().collect(Collectors.toList());
+                for(int i = 0; i < list.size(); i++) {
+                    String varType = list.get(i);
+                    int nbVariables = mapTypeClassNbVars.get(c).get(varType).size();
+                    long nbInstantiatedVariables = mapTypeClassNbVars.get(c).get(varType).stream().filter(Variable::isInstantiated).count();
+                    long nbConstantVariables = mapTypeClassNbVars.get(c).get(varType).stream().filter(Variable::isAConstant).count();
+                    long nbViews = mapTypeClassNbVars.get(c).get(varType).stream().filter(var -> (var.getTypeAndKind() & Variable.VIEW) != 0).count();
+                    if (mapTypeClassNbVars.get(c).get(varType).stream().allMatch(IView.class::isInstance)) {
+                        nbViews = 0;
+                    }
[…]
