### 2022-10-05 810e90e
Few corrections and user-oriented improvements
