+++ paper/paper.md
@@ -42,19 +42,13 @@ The implementation of new constraints [@10.1007/978-3-031-08011-1_21]
 or strategies for exploring the search space [@DBLP:conf/cp/LiYL21;@fages:hal-01629182] 
 is therefore possible.
 
-As a result, `Choco-solver` is used by the academy for teaching and research and by the industry to solve real-world problems, 
-such as cryptanalysis [@10.1007/978-3-030-78375-4_8],
-planing construction [@CANIZARES2022116149], 
-automated testing and debugging [@LE2021100085],
-scheduling [@10.1007/978-3-319-44953-1_40],
-level design [@5887401]
-placement service [@8814186] and many others.
+As a result, `Choco-solver` is used by the academy for teaching and research and by the industry to solve real-world problems.
              
 ## CP in a nutshell
-Like integer linear programming or Boolean satisfaction, constraint programming is a field of mathematical programming.
-It focuses on describing and solving applied mathematical problems.
-However, what distinguishes the CP from the first two approaches is that it relies on a high level language to 
-describe the problems. 
+Constraint programming provides not only a declarative way for user to describe discrete problems,
+but also technics for solving them automatically.  
+In that sense, it is very close to integer linear programming or Boolean satisfaction
+but is distinguished from them with its high-level modeling language and expressiveness.
 Actually, one of the richness's of the paradigm lies in the wide variety of constraints 
 it proposes, which are also central to the solving stage. 
 Thus, the objective of CP is twofold: firstly to offer a rich declarative language to describe a combinatorial problem, 
@@ -62,7 +56,7 @@ and secondly to provide technics to automatically solve this problem.
 In standard use, a user states a problem using variables, their domains (possible values for each
 variable), and constraints, predicates that must hold on the variables.
 The wide variety of constraints available allows the user to describe his problem as naturally as possible.
-Each constraint ensures that it holds, otherwise a propagator *filters* ,removes, 
+Each constraint ensures that it holds, otherwise a propagator filters  
 from the domain of variables the values that prevent the satisfiability.
 It is the combination of the selected constraints that defines the problem to be solved.
 The problem is solved by alternating space reduction (usually by a depth-first search) and propagation, 
@@ -92,17 +86,17 @@ into both academic and industrial projects.
 ## Modeling                                                        
 
 `Choco-solver` comes with the commonly used types of variables: 
-integer variables (with either bounded or enumerated domain), 
+integer variables, 
 Boolean variables, set variables [@gervet_1997] graph variables [@dooms_2005;@fages_2015] and real variables. 
 Views [@DBLP:conf/cp/SchulteT05;@DBLP:journals/constraints/Justeau-Allaire22] 
 but also arithmetical, relational and logical expressions are supported.
 
-Up to 100 constraints --and more than 150 propagators-- are provided : 
+Up to 100 constraints are provided : 
 from classic ones, such as arithmetical constraints, 
 to  must-have global constraints, such as *AllDifferent* [@10.5555/199288.178024] or *Cumulative* [@aggoun:hal-00442821],
 and include less common even though useful ones, such as  *Tree* [@DBLP:conf/cpaior/BeldiceanuFL05] 
 or *StableKeySort* [@beldiceanu:hal-01186680].
-In many cases, the Choco-Solver API provides various options in addition to the default signature - 
+In many cases, the `Choco-solver` API provides various options in addition to the default signature - 
 corresponding to a robust implementation – of a constraint. 
 This allows the user to experiment alternative approaches and tune the model to its instance. 
[…]
