+++ paper/paper.bib
@@ -532,3 +532,17 @@ abstract="The cumulative constraint is the key to the success of Constraint Prog
 isbn="978-3-031-08011-1",
 doi={10.1007/978-3-031-08011-1_21}
 }
+@online{OptaForce,
+  author="Fabienne Grazzini",
+  title="Airbus and COSLING provide software solution Optaforce for Mirage 2000 maintenance",
+  year="2019",
+  month="June",
+  url="https://www.airbus.com/en/newsroom/press-releases/2019-06-airbus-and-cosling-provide-software-solution-optaforce-for-mirage"
+}
+
+@inproceedings{Charpentier2021COSLINGC,
+  title={COSLING Configurator},
+  author={Antoine Charpentier and Jean-Guillaume Fages and Tanguy Lap{\`e}gue},
+  booktitle={ConfWS},
+  year={2021}
+}
