### 2022-10-05 #4708 @cprudhom
I just pushed a new version of the article, with (mainly) a new paragraph on industrial applications.
I also conclude the `Features and Functionality` paragraph with few words on other Java solvers (but to be clear, I'm not really comfortable with this). I though it was easier for me to write this here instead of `Statement of need` because it mentioned concepts/functionalities presented after.

Best regards
