### 2022-10-05 #4708 @cprudhom
Thinking twice, I'm not satisfied with the comparison with the (two) other solvers.
Even if I could have made it deeper, it certainly requires a fair comparison with other CSP solvers, Java or not, with a commercial licence or not. But, I believe this is out of the scope.
What do you think?
