### 2022-10-06 #4708 @skadio
I found the Community/Integration section quite useful and a nice addition. It will definitely be useful for readers to learn about other relevant technology such as Minizinc/XCSP etc. The quick commentary on other (java) solvers is fair/reasonable.

I agree that going through an entire listing of CP solvers/tools would be out of scope, but at the very least, covering the "java" part seems relevant to me, since that's an angle of the paper as well.

There is an item in JOSS Review Checklikst:
* Do the authors describe how this software compares to other commonly-used packages?

This new section checks that box in my mind.
