+++ paper/paper.bib
@@ -366,7 +366,7 @@ isbn="978-3-540-30201-8"
 }
 @inproceedings{allouche:hal-01198361,
   TITLE = {{Anytime hybrid best-first search with tree decomposition for weighted CSP}},
-  AUTHOR = {Allouche, David and De Givry, Simon and KATSIRELOS, Georgios and Schiex, Thomas and Zytnicki, Matthias},
+  AUTHOR = {Allouche, David and De Givry, Simon and Katsirelos, Georgios and Schiex, Thomas and Zytnicki, Matthias},
   URL = {https://hal.archives-ouvertes.fr/hal-01198361},
   BOOKTITLE = {{CP 2015 - 21st International Conference on Principles and Practice of Constraint Programming}},
   ADDRESS = {Cork, Ireland},
