+++ paper/paper.md
@@ -27,14 +27,14 @@ bibliography: paper.bib
 
 Constraint Programming (CP) is a powerful programming paradigm for solving 
 combinatorial search problems [@DBLP:reference/fai/2].
-CP is at the intersection of artificial intelligence, computer science, operations research and many other fields.
+CP is at the intersection of artificial intelligence, computer science, operations research, and many other fields.
 One of the richness's of the paradigm lies in the wide variety of constraints it proposes. 
 Thus, the benefit of CP is twofold: firstly to offer a rich declarative language to describe a combinatorial problem, 
-and secondly to provide technics to automatically solve this problem.
+and secondly to provide techniques to automatically solve this problem.
 
 `Choco-solver` is Java library for constraint programming which was created in the early 2000s.
 Since then, the library has evolved a great deal, but ease of use has always been a guiding principle in its development.
-The `Choco-solver` API is designed to reduce entry points to a minimum and thus simplify modelling for users.
+The `Choco-solver` API is designed to reduce entry points to a minimum and thus simplifies modelling for users.
 The wide variety of constraints available allows the user to describe his problem as naturally as possible.
 The black-box approach to solving allows everyone to focus on modelling.
 However, `Choco-solver` is also open and modifiable.
@@ -42,38 +42,38 @@ The implementation of new constraints [@10.1007/978-3-031-08011-1_21]
 or strategies for exploring the search space [@DBLP:conf/cp/LiYL21;@fages:hal-01629182] 
 is therefore possible.
 
-As a result, `Choco-solver` is used by the academy for teaching and research and by the industry to solve real-world problems.
+As a result, `Choco-solver` is used by the academics for teaching and research, on the other hand it is used by the industry to solve real-world problems.
              
 ## CP in a nutshell
-Constraint programming provides not only a declarative way for user to describe discrete problems,
-but also technics for solving them automatically.  
+Constraint programming provides not only a declarative way for users to describe discrete problems,
+but also technique for solving them automatically.  
 In that sense, it is very close to integer linear programming or Boolean satisfaction
 but is distinguished from them with its high-level modeling language and expressiveness.
 Actually, one of the richness's of the paradigm lies in the wide variety of constraints 
 it proposes, which are also central to the solving stage. 
 Thus, the objective of CP is twofold: firstly to offer a rich declarative language to describe a combinatorial problem, 
-and secondly to provide technics to automatically solve this problem.
+and secondly to provide technics for solving the problem automatically.
 In standard use, a user states a problem using variables, their domains (possible values for each
-variable), and constraints, predicates that must hold on the variables.
-The wide variety of constraints available allows the user to describe his problem as naturally as possible.
+variable), and constraints which are called predicates that must hold on the variables.
+The wide variety of constraints available allows the users to describe their problem as naturally as possible.
 Each constraint ensures that it holds, otherwise a propagator filters  
-from the domain of variables the values that prevent the satisfiability.
+the values that prevent the satisfiability.
 It is the combination of the selected constraints that defines the problem to be solved.
 The problem is solved by alternating space reduction (usually by a depth-first search) and propagation, 
 thus ensuring the completeness of the approach.
 This standard usage can be extended in different ways, for example, by hybridisation with local search, 
-Boolean satisfiability or linear programming techniques.
+Boolean satisfiability, or linear programming techniques.
 
 # Statement of need
[…]
