### 2022-10-10 #4708 @cprudhom
@jbytecode I accepted the PR, thank you.
Last question: how easy is it the upgrade the paper in the future ?
For instance, we are working on Python bindings for Choco, which would be interesting to indicate in the paper.
Should we delay the current process or restart it when it is ready?
