### 2022-10-11 26410c5
Few corrections on paper.md
