### 2022-10-11 #4708 @cprudhom
@jbytecode

> Create a tagged release in the software repository ... The version should be in form of vx.y.z, e.g. v1.2.3.

Our release versioning system is based on Maven and does not show the `v`. Is this necessary?
