+++ CHANGES.md
@@ -3,7 +3,20 @@ Choco Solver ChangeLog
 
 This file is dedicated to sum up the new features added and bugs fixed in Choco-solver since the version, 4.0.0.
 **Note**: double-space is replaced by "\t" character on release process. Make sure the format is ok.
-    
+
+NEXT MILESTONE
+-------------------
+
+### Major features:
+
+### Deprecated API (to be removed in next release):
+
+### Other closed issues and pull requests:
+See [milestone 4.10.11](https://github.com/chocoteam/choco-solver/milestone/xx)
+
+#### Contributors to this release:
+
+        
 4.10.10 - 11 Oct 2022
 -------------------
 
