### 2022-10-11 #4708 @cprudhom
@jbytecode

I believe everything is ready:
- tagged release : [v4.10.10](https://github.com/chocoteam/choco-solver/releases/tag/v4.10.10)
- Zenodo archive DOI : 10.5281/zenodo.7185962
- Zenodo archive URL : https://zenodo.org/record/7185962#.Y0WCVuxBxjw
