+++ paper/paper.md
@@ -15,6 +15,7 @@ authors:
     affiliation: 2
 affiliations:
   - name: TASC, IMT-Atlantique, LS2N-CNRS, Nantes, France
+    orcid: 0000-0001-5638-4777
     index: 1
   - name: COSLING S.A.S., Nantes, France
     index: 2
