### 2022-10-11 #4708 @jbytecode
@cprudhom - It seems the ORCID of the co-author is missing the paper. Please merge the pull request: https://github.com/chocoteam/choco-solver/pull/947
