+++ solver/src/main/java/org/chocosolver/solver/constraints/IGraphConstraintFactory.java
@@ -61,7 +61,7 @@ public interface IGraphConstraintFactory extends ISelf<Model> {
      * @param nb an integer variable indicating the expected number of nodes in g
      * @return A constraint to force the number of nodes in g to be equal to nb
      */
-    default Constraint nbNodes(GraphVar g, IntVar nb) {
+    default Constraint nbNodes(GraphVar<?> g, IntVar nb) {
         return new Constraint("nbNodes", new PropNbNodes(g, nb));
     }
 
@@ -72,7 +72,7 @@ public interface IGraphConstraintFactory extends ISelf<Model> {
      * @param nb an integer variable indicating the expected number of edges in g
      * @return A constraint to force the number of edges in g to be equal to nb
      */
-    default Constraint nbEdges(GraphVar g, IntVar nb) {
+    default Constraint nbEdges(GraphVar<?> g, IntVar nb) {
         return new Constraint("nbEdges", new PropNbEdges(g, nb));
     }
 
@@ -86,7 +86,7 @@ public interface IGraphConstraintFactory extends ISelf<Model> {
      * @param g a graph variable
      * @return A constraint which makes sure every node has a loop
      */
-    default Constraint loopSet(GraphVar g, SetVar loops) {
+    default Constraint loopSet(GraphVar<?> g, SetVar loops) {
         return new Constraint("loopSet", new PropLoopSet(g, loops));
     }
 
@@ -98,7 +98,7 @@ public interface IGraphConstraintFactory extends ISelf<Model> {
      * @param nb an integer variable counting the number of loops in g
      * @return A constraint which ensures g has nb loops
      */
-    default Constraint nbLoops(GraphVar g, IntVar nb) {
+    default Constraint nbLoops(GraphVar<?> g, IntVar nb) {
         return new Constraint("nbLoops", new PropNbLoops(g, nb));
     }
 
@@ -111,23 +111,23 @@ public interface IGraphConstraintFactory extends ISelf<Model> {
     // symmetry
 
     /**
-     * Creates a constraint which ensures that g is a symmetric directed graph
+     * Creates a constraint which ensures that g is a symetric directed graph
      * This means (i,j) in g <=> (j,i) in g
      * Note that it may be preferable to use an undirected graph variable instead!
      *
      * @param g a directed graph variable
-     * @return A constraint which ensures that g is a symmetric directed graph
+     * @return A constraint which ensures that g is a symetric directed graph
      */
     default Constraint symmetric(DirectedGraphVar g) {
         return new Constraint("symmetric", new PropSymmetric(g));
     }
 
     /**
-     * Creates a constraint which ensures that g is an antisymmetric directed graph
+     * Creates a constraint which ensures that g is an antisymetric directed graph
      * This means (i,j) in g => (j,i) notin g
      *
      * @param g a directed graph variable
-     * @return A constraint which ensures that g is an antisymmetric directed graph
+     * @return A constraint which ensures that g is an antisymetric directed graph
      */
     default Constraint antisymmetric(DirectedGraphVar g) {
         return new Constraint("antisymmetric", new PropAntiSymmetric(g));
@@ -144,8 +144,8 @@ public interface IGraphConstraintFactory extends ISelf<Model> {
      * @param g A graph variable
      * @return A transitivity constraint
      */
-    default Constraint transitivity(GraphVar g) {
-        return new Constraint("transitivity", new PropTransitivity(g));
+    default Constraint transitivity(GraphVar<?> g) {
+        return new Constraint("transitivity", new PropTransitivity<>(g));
     }
 
[…]
