### 2022-10-11 #4708 @cprudhom
> @cprudhom - It seems the ORCID of the co-author is missing in the paper. Please merge the pull request: [chocoteam/choco-solver#947](https://github.com/chocoteam/choco-solver/pull/947)

Right, thank you. Does it require a new release?
