### 2022-10-11 #4708 @jbytecode
@cprudhom - no, the release and archive is okay.

I am re-generating the pdf, please have a proof read and ping me again.
