### 2022-10-12 #4708 @cprudhom
@jbytecode I think the sub-section `Applications` is useless. I can fix that if you want, otherwise, I'm ok with the article
