### 2022-10-12 #4708 @jbytecode
@cprudhom - the review has been completed so it is usual to go on with the typo and language corrections. I am now recommending an accept. Thank you for your edits.

@skadio, @ozgurakgun - Thank you so much again, for your great effort and consuming your time for the JOSS review, hope to work on future projects.
