### 2022-10-12 #4708 @jbytecode
@cprudhom - interestingly the orchid is not shown properly, could you please solve this?
