### 2022-10-12 c3a5527
Work on objective manager: - reduce usages of `Number` when not necessary, for efficiency reason  - remove AbstractObjManager.java
