### 2022-10-12 a44be52
reason: '<>' with anonymous inner classes is not supported in -source 8
