### 2022-10-13 #4708 @cprudhom
Thank you @jbytecode Do you know whether such a procedure is possible on https://joss.theoj.org/papers/10.21105/joss.04708 too?
