### 2022-10-21 #4708 @xuanxu
> @xuanxu Do I need to open a specific issue for the tag problem?

No need, I'm going to take a look at how to best update that metadata
